pythonTips/requirements.txt
openai==1.9.0
openpyxl==3.1.2
python-dotenv==1.0.1
requests==2.31.0
twilio==8.11.1
